refactor(scoring): move PeerScoringService into core
The reference scoring service is protocol policy, not infrastructure —
it has no I/O and is just storage-agnostic state-machine logic. Storage
backend (InMemoryPeerScoreStorage) and delta-table provider
(FixedScoringProvider) stay in app as swappable choices.
