+++ src/core/peer_scoring.rs
@@ -1,7 +1,8 @@
-//! Types and traits core uses to describe scoring-relevant events plus
-//! pure score-derivation helpers. The scoring service itself lives in
-//! `crate::app::peer_scoring` and consumes the `Vec<ScoreOp>` that the
-//! helpers here return.
+//! Peer-scoring vocabulary, traits, reference [`PeerScoringService`],
+//! and pure score-derivation helpers. The service is storage- and
+//! policy-agnostic; concrete backends (e.g.
+//! [`crate::app::InMemoryPeerScoreStorage`]) and delta-table providers
+//! (e.g. [`crate::app::FixedScoringProvider`]) live in the app layer.
 
 use prost::Message;
 
@@ -92,8 +93,7 @@ pub struct ScoringMemberDiff {
 }
 
 /// Pure diff between a scoring table snapshot and an MLS member roster.
-/// Caller applies the diff to its own
-/// [`PeerScoringService`](crate::app::PeerScoringService).
+/// Caller applies the diff to its own [`PeerScoringService`].
 pub fn scoring_member_diff(scored: &[Vec<u8>], mls_members: &[Vec<u8>]) -> ScoringMemberDiff {
     let scored_set: std::collections::HashSet<&[u8]> = scored.iter().map(Vec::as_slice).collect();
     let mls_set: std::collections::HashSet<&[u8]> = mls_members.iter().map(Vec::as_slice).collect();
@@ -155,10 +155,285 @@ fn creator_penalty(ev: &ViolationEvidence) -> ScoreOp {
     }
 }
 
+// ── Reference scoring service ───────────────────────────────────────
+
+/// Per-group, per-member score tracker. One instance per group;
+/// threshold travels with [`ScoringConfig`]. Storage is abstracted via
+/// [`PeerScoreStorage`] so app-layer backends (in-memory, on-disk, …)
+/// plug in without touching this protocol logic.
+pub struct PeerScoringService<S: PeerScoreStorage, P: ScoringProvider> {
+    storage: S,
+    provider: P,
+    config: ScoringConfig,
+}
+
+impl<S: PeerScoreStorage, P: ScoringProvider> PeerScoringService<S, P> {
+    pub fn new(storage: S, provider: P, config: ScoringConfig) -> Self {
+        Self {
+            storage,
+            provider,
+            config,
+        }
+    }
+
+    /// Start tracking a member with the default score.
+    pub fn add_member(&mut self, member_id: &[u8]) {
+        self.storage.set(member_id, self.config.default_score);
+    }
+
+    pub fn remove_member(&mut self, member_id: &[u8]) {
+        self.storage.remove(member_id);
+    }
+
+    /// Apply a single [`ScoreOp`] and return the target's new score, or
+    /// `None` if the target isn't tracked.
+    pub fn apply_op(&mut self, op: &ScoreOp) -> Option<i64> {
+        let current = self.storage.get(&op.member_id)?;
+        let delta = self.provider.score_delta(op.event);
+        let new_score = current.saturating_add(delta);
+        self.storage.set(&op.member_id, new_score);
+        Some(new_score)
+    }
+
+    /// Apply a batch of [`ScoreOp`]s. Untracked targets are skipped
+    /// silently (same semantics as [`Self::apply_op`]).
+    pub fn apply_ops(&mut self, ops: &[ScoreOp]) {
+        for op in ops {
+            self.apply_op(op);
+        }
+    }
+
+    pub fn score_for(&self, member_id: &[u8]) -> Option<i64> {
+        self.storage.get(member_id)
+    }
+
+    /// Force-set a score; used when applying a `GroupSync` from the steward.
+    pub fn set_score(&mut self, member_id: &[u8], score: i64) {
+        self.storage.set(member_id, score);
+    }
+
+    pub fn members_below_threshold(&self) -> Vec<Vec<u8>> {
+        let threshold = self.config.threshold;
+        self.storage
+            .all_scores()
+            .into_iter()
+            .filter(|(_, score)| *score <= threshold)
+            .map(|(id, _)| id)
+            .collect()
+    }
+
+    pub fn is_below_threshold(&self, member_id: &[u8]) -> bool {
+        self.storage
+            .get(member_id)
+            .is_some_and(|s| s <= self.config.threshold)
+    }
+
+    pub fn all_members_with_scores(&self) -> Vec<(Vec<u8>, i64)> {
+        self.storage.all_scores()
+    }
+
+    pub fn config(&self) -> &ScoringConfig {
+        &self.config
+    }
+
+    pub fn set_threshold(&mut self, threshold: i64) {
+        self.config.threshold = threshold;
+    }
+}
+
 #[cfg(test)]
 mod tests {
+    use std::collections::HashMap;
+
     use super::*;
     use crate::protos::de_mls::messages::v1::{EmergencyCriteriaProposal, ViolationType};
+
+    // ── Test scaffolding ────────────────────────────────────────────
+
+    /// Minimal in-memory storage for service tests. Production storage
+    /// lives in [`crate::app::InMemoryPeerScoreStorage`].
+    #[derive(Default)]
+    struct TestStorage(HashMap<Vec<u8>, i64>);
+
+    impl PeerScoreStorage for TestStorage {
+        fn get(&self, member_id: &[u8]) -> Option<i64> {
+            self.0.get(member_id).copied()
+        }
+        fn set(&mut self, member_id: &[u8], score: i64) {
+            self.0.insert(member_id.to_vec(), score);
+        }
+        fn remove(&mut self, member_id: &[u8]) {
+            self.0.remove(member_id);
+        }
+        fn all_scores(&self) -> Vec<(Vec<u8>, i64)> {
+            self.0.iter().map(|(k, v)| (k.clone(), *v)).collect()
+        }
+    }
+
+    /// HashMap-backed [`ScoringProvider`] with caller-supplied deltas.
+    /// Production deltas live in [`crate::app::FixedScoringProvider`].
+    struct TestProvider(HashMap<ScoreEvent, i64>);
+
+    impl ScoringProvider for TestProvider {
+        fn score_delta(&self, event: ScoreEvent) -> i64 {
+            self.0.get(&event).copied().unwrap_or(0)
+        }
+    }
+
+    fn make_service() -> PeerScoringService<TestStorage, TestProvider> {
+        let deltas = HashMap::from([
+            (ScoreEvent::EmergencyNoCreator, -50),
+            (ScoreEvent::EmergencyYesCreator, 20),
+            (ScoreEvent::BrokenCommit, -50),
+            (ScoreEvent::SuccessfulCommit, 10),
+            (ScoreEvent::MisbehavingCommit, -30),
+        ]);
+        PeerScoringService::new(
+            TestStorage::default(),
+            TestProvider(deltas),
+            ScoringConfig {
+                default_score: 100,
+                threshold: 0,
+            },
+        )
+    }
+
+    // ── Service tests ────────────────────────────────────────────────
+
+    #[test]
+    fn add_member_gets_default_score() {
+        let mut svc = make_service();
+        svc.add_member(b"alice");
+        assert_eq!(svc.score_for(b"alice"), Some(100));
+    }
+
+    #[test]
+    fn unknown_member_returns_none() {
+        let svc = make_service();
+        assert_eq!(svc.score_for(b"unknown"), None);
+    }
+
+    #[test]
+    fn remove_member_clears_score() {
+        let mut svc = make_service();
+        svc.add_member(b"alice");
+        svc.remove_member(b"alice");
+        assert_eq!(svc.score_for(b"alice"), None);
+    }
+
+    #[test]
+    fn apply_event_decreases_score() {
+        let mut svc = make_service();
+        svc.add_member(b"alice");
+        let new_score = svc.apply_op(&ScoreOp {
+            member_id: b"alice".to_vec(),
+            event: ScoreEvent::EmergencyNoCreator,
+        });
+        assert_eq!(new_score, Some(50));
+        assert_eq!(svc.score_for(b"alice"), Some(50));
+    }
+
+    #[test]
+    fn apply_event_unknown_member_returns_none() {
+        let mut svc = make_service();
+        let result = svc.apply_op(&ScoreOp {
+            member_id: b"unknown".to_vec(),
+            event: ScoreEvent::EmergencyNoCreator,
+        });
+        assert_eq!(result, None);
+    }
+
+    #[test]
+    fn multiple_events_accumulate() {
+        let mut svc = make_service();
+        svc.add_member(b"alice");
+        for event in [
+            ScoreEvent::EmergencyNoCreator,
+            ScoreEvent::MisbehavingCommit,
+            ScoreEvent::SuccessfulCommit,
+        ] {
+            svc.apply_op(&ScoreOp {
+                member_id: b"alice".to_vec(),
+                event,
+            });
+        }
+        assert_eq!(svc.score_for(b"alice"), Some(30));
+    }
+
+    #[test]
+    fn members_below_threshold_filters_correctly() {
+        let mut svc = make_service();
+        svc.add_member(b"alice");
+        svc.add_member(b"bob");
+        svc.add_member(b"charlie");
+        for event in [ScoreEvent::EmergencyNoCreator, ScoreEvent::BrokenCommit] {
+            svc.apply_op(&ScoreOp {
+                member_id: b"alice".to_vec(),
+                event,
+            });
+        }
+        for _ in 0..2 {
+            svc.apply_op(&ScoreOp {
+                member_id: b"charlie".to_vec(),
+                event: ScoreEvent::EmergencyNoCreator,
+            });
+        }
+        let below = svc.members_below_threshold();
+        assert!(below.contains(&b"alice".to_vec()));
+        assert!(below.contains(&b"charlie".to_vec()));
+        assert!(!below.contains(&b"bob".to_vec()));
+    }
+
+    #[test]
+    fn set_threshold_changes_below_threshold_set() {
+        let mut svc = make_service();
+        svc.add_member(b"alice");
+        svc.set_score(b"alice", -10);
+
+        svc.set_threshold(-50);
+        assert!(!svc.members_below_threshold().contains(&b"alice".to_vec()));
+
+        svc.set_threshold(-5);
+        assert!(svc.members_below_threshold().contains(&b"alice".to_vec()));
+    }
+
+    #[test]
+    fn score_saturates_no_overflow() {
+        let mut svc = PeerScoringService::new(
+            TestStorage::default(),
+            TestProvider(HashMap::from([(ScoreEvent::SuccessfulCommit, i64::MAX)])),
+            ScoringConfig {
+                default_score: i64::MAX,
+                threshold: 0,
+            },
+        );
+        svc.add_member(b"alice");
+        let new_score = svc.apply_op(&ScoreOp {
+            member_id: b"alice".to_vec(),
+            event: ScoreEvent::SuccessfulCommit,
+        });
+        assert_eq!(new_score, Some(i64::MAX));
+    }
+
+    #[test]
+    fn unknown_event_yields_zero_delta() {
+        let mut svc = PeerScoringService::new(
+            TestStorage::default(),
+            TestProvider(HashMap::from([(ScoreEvent::EmergencyNoCreator, -50)])),
+            ScoringConfig {
+                default_score: 100,
+                threshold: 0,
+            },
+        );
+        svc.add_member(b"alice");
+        let new_score = svc.apply_op(&ScoreOp {
+            member_id: b"alice".to_vec(),
+            event: ScoreEvent::SuccessfulCommit,
+        });
+        assert_eq!(new_score, Some(100));
+    }
+
+    // ── ECP score derivation tests ──────────────────────────────────
 
     fn ecp_payload(violation_type: i32, target: Vec<u8>, creator: Vec<u8>) -> Vec<u8> {
         let evidence = ViolationEvidence {
